// apps/admin/src/api/generated/schemas/appointmentRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { AppointmentReadClinicianId } from './appointmentReadClinicianId';
import type { AppointmentReadBranchId } from './appointmentReadBranchId';
import type { AppointmentType } from './appointmentType';
import type { AppointmentStatus } from './appointmentStatus';
import type { AppointmentReadNotes } from './appointmentReadNotes';
import type { AppointmentReadPartyName } from './appointmentReadPartyName';
import type { AppointmentReadTenantName } from './appointmentReadTenantName';
import type { AppointmentReadType } from './appointmentReadType';

export interface AppointmentRead {
  /** Unique identifier for the resource */
  id: string;
  partyId: string;
  clinicianId?: AppointmentReadClinicianId;
  branchId?: AppointmentReadBranchId;
  date: string;
  /** @pattern ^\d{2}:\d{2}$ */
  time: string;
  duration?: number;
  appointmentType?: AppointmentType;
  status?: AppointmentStatus;
  notes?: AppointmentReadNotes;
  tenantId: string;
  partyName?: AppointmentReadPartyName;
  tenantName?: AppointmentReadTenantName;
  type?: AppointmentReadType;
}